:recycle: refactor profile page and view profile page into one
